+++ packages/phoenix/src/components/Dropdown/Internal/DropdownItem.vue
@@ -1,70 +1,48 @@
 <script setup lang="ts">
-import { computed } from 'vue';
+import { onMounted, ref, watch } from 'vue';
 import type { DropdownItemProps } from '~/types';
-import { Checkbox, Thumbnail } from '~/components';
+import { Checkbox } from '~/components';
 
-const props = withDefaults(
-  defineProps<DropdownItemProps>(),
-  { checkbox: false },
-);
+const props = defineProps<DropdownItemProps>();
 
-const emit = defineEmits(['toggle']);
+const emit = defineEmits(['onSelect']);
 
-const model = computed<boolean>({
-  get: () => props.selected,
-  set: (value: boolean) => emit('toggle', value),
+const checked = ref(false);
+
+onMounted(() => {
+  checked.value = props.selected;
 });
 
-function toggle(event: Event) {
-  event.preventDefault();
-  model.value = !model.value;
-}
+watch(() => props.selected, (newValue) => {
+  checked.value = newValue;
+});
 </script>
 
 <template>
-  <button role="button" class="item" :class="{ selected }" @click="(event) => toggle(event)">
-    <Checkbox v-if="checkbox" v-model="model" tabindex="-1" class="checkbox" />
-    <div v-if="item.image || item.icon" class="preview">
-      <Thumbnail v-if="item.image" size="small" :src="item.image" :alt="item.label" class="preview-image" />
-      <i class="preview-icon" :class="[item.icon]" />
-    </div>
-    <div class="label">
-      {{ item.label }}
-    </div>
-    <slot />
-  </button>
+  <div class="dropdown-item" :class="[{ selected }]" @click="() => emit('onSelect')">
+    <Checkbox v-if="multiple" v-model="checked" />
+    <slot>
+      <div class="label">
+        {{ item.label }}
+      </div>
+    </slot>
+  </div>
 </template>
 
 <style scoped>
-.item {
+.dropdown-item {
   display: flex;
   align-items: center;
-  width: 100%;
-  padding: 12px 16px;
-  border: none;
-  background-color: var(--base-white);
-  cursor: pointer;
-  user-select: none;
-  gap: 8px;
-}
-
-.item .label {
-  color: var(--gray-900);
+  padding: 16px;
+  transition: background-color 150ms linear;
   font: var(--text-sm-regular);
+  cursor: pointer;
+  gap: 16px;
 }
 
-.item:focus,
-.item:active {
-  outline: none;
-  background: var(--gray-100);
-}
-
-.item.selected,
-.item:hover {
+.dropdown-item.selected,
+.dropdown-item:hover,
+.dropdown-item:focus {
   background-color: var(--gray-50);
-}
-
-.preview {
-  color: var(--gray-500);
 }
 </style>
