--- packages/phoenix/src/components/DropdownV3/type.ts
@@ -1,21 +1,0 @@
-export interface DropdownItemType {
-  label: string
-  key: string | number
-  groupName?: string
-}
-
-export interface DropdownItemProps {
-  item: DropdownItemType
-  multiple?: boolean
-  selected?: boolean
-}
-
-export interface DropdownProps {
-  modelValue: DropdownItemType | DropdownItemType[] | null
-  placeholder?: string
-  items: DropdownItemType[]
-  multiple?: boolean
-  multiSelectLabel?: string
-  isLoading?: boolean
-  noDataText?: string
-}
