+++ packages/phoenix/src/components/index.ts
@@ -1,5 +1,4 @@
 export { default as Dropdown } from './Dropdown/Dropdown.vue';
-export { default as DropdownList } from './Dropdown/DropdownList.vue';
 export { default as Note } from './Note/Note.vue';
 export { default as Progress } from './Progress/Progress.vue';
 export { default as Checkbox } from './Checkbox/Checkbox.vue';
@@ -19,13 +18,11 @@ export { default as MediaInput } from './Upload/MediaInput.vue';
 export { default as UploadedMedia } from './Upload/UploadedMedia.vue';
 export { default as Increment } from './Increment/Increment.vue';
 export { default as Tag } from './Tag/Tag.vue';
-export { default as MultiselectDropdown } from './Dropdown/MultiselectDropdown.vue';
 export { default as Status } from './Status/Status.vue';
 export { default as DatePicker } from './DateInput/DatePicker.vue';
 export { default as DateInput } from './DateInput/DateInput.vue';
 export { default as SingleDatePicker } from './DateInput/SingleDatePicker.vue';
 export { default as SingleDateInput } from './DateInput/SingleDateInput.vue';
-export { default as DropdownPrefix } from './Input/prefixes/DropdownPrefix.vue';
 export { default as Tab } from './Tabs/Tab.vue';
 export { default as TabsBar } from './Tabs/TabsBar.vue';
 export { default as Sidebar } from './Sidebar/Sidebar.vue';
